clean up implementation
Removed NodeBuilder, AbstractNodeBuilder, ConnectedGraph and
TrickleGraph and replaced all of that with GraphBuilder and
PreparedGraph.

Changed the publicly exposed interfaces to abstract classes with package
private constructors. This was needed in order to allow for a package
private run method that continues running on an existing state. This is
needed when following dependencies into other graph instances.